modify append dependency properties function

+++ src/main/java/com/tzg/plugin/dependency/support/MongoDBSupport.java
@@ -12,7 +12,7 @@ import java.util.Properties;
 
 public final class MongoDBSupport {
 
-    public static int MONGODB_COMMENT_LENGTH = 11;
+    public static int MONGODB_COMMENT_LENGTH = 12;
 
     public static Map< String, String > getMongoDBMap() {
 
@@ -42,12 +42,13 @@ public final class MongoDBSupport {
         sb.append( "\n## mongoDB config:                                                                          ##" );
         sb.append( "\n##                                                                                          ##" );
         sb.append( "\n## connections-per-host: 每个主机答应的连接数(每个主机的连接池大小), 当连接池被用光时, 会被阻塞住   ##" );
-        sb.append( "\n## max-wait-time:        被阻塞线程从连接池获取连接的最长等待时间（ms)                           ##" );
+        sb.append( "\n## max-wait-time:        被阻塞线程从连接池获取连接的最长等待时间(ms)                            ##" );
         sb.append( "\n## connect-timeout:      在建立(打开)套接字连接时的超时时间(ms)                                 ##" );
-        sb.append( "\n## socket-timeout:       套接字超时时间; 该值会被传递给Socket.setSoTimeout（int）               ##" );
+        sb.append( "\n## socket-timeout:       套接字超时时间; 该值会被传递给Socket.setSoTimeout(int)                ##" );
         sb.append( "\n## slave-ok:             指明是否答应驱动从次要节点或者slave节点读取数据                         ##" );
         sb.append( "\n##                                                                                          ##" );
         sb.append( "\n## ---------------------------------------------------------------------------------------- ##" );
+        sb.append( "\n" );
         return sb.toString();
     }
 
@@ -64,7 +65,9 @@ public final class MongoDBSupport {
 
     public static void removeMongoDBModule() throws IOException {
         File file = new File( getMongoDBModulePath() );
-        FileUtils.forceDelete( file );
+        if ( file.exists() ) {
+            FileUtils.forceDelete( file );
+        }
         System.out.println( "====> delete mongodb module '$path' successfully.".replace( "$path", file.getAbsolutePath() ) );
     }
